Revert "Feat: Record serialization/deserialization (#1706)"
This reverts commit f5613b4ecb46c61cd7c24ae55606a63f92303360.
