+++ google-cloud-firestore/src/main/java/com/google/cloud/firestore/DocumentSnapshot.java
@@ -19,7 +19,6 @@ package com.google.cloud.firestore;
 import com.google.api.core.InternalExtensionOnly;
 import com.google.cloud.Timestamp;
 import com.google.cloud.firestore.UserDataConverter.EncodingOptions;
-import com.google.cloud.firestore.encoding.CustomClassMapper;
 import com.google.common.base.Preconditions;
 import com.google.firestore.v1.Document;
 import com.google.firestore.v1.Value;
