+++ google-cloud-firestore/src/main/java/com/google/cloud/firestore/FieldValue.java
@@ -16,7 +16,6 @@
 
 package com.google.cloud.firestore;
 
-import com.google.cloud.firestore.encoding.CustomClassMapper;
 import com.google.common.base.Preconditions;
 import com.google.firestore.v1.ArrayValue;
 import com.google.firestore.v1.DocumentTransform.FieldTransform;
