+++ google-cloud-firestore/src/test/java/com/google/cloud/firestore/ToStringTest.java
@@ -19,7 +19,6 @@ package com.google.cloud.firestore;
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.cloud.Timestamp;
-import com.google.cloud.firestore.encoding.CustomClassMapper;
 import com.google.cloud.firestore.spi.v1.FirestoreRpc;
 import com.google.firestore.v1.Value;
 import java.util.Collections;
